microservices/microservice-2/requirements.txt
boto3==1.18.65
python-dotenv==0.19.0